+++ pubmodules/trigger/rulesTrigger.js
@@ -1192,7 +1192,6 @@ class RulesTrigger {
                       lead: createdLead, requester: puser
                     };
       
-
                     return requestService.create(new_request).then(function (savedRequest) {                   
 
                         if (attributes) {