+++ routes/kb.js
@@ -30,6 +30,11 @@ const KB_WEBHOOK_TOKEN = process.env.KB_WEBHOOK_TOKEN || 'kbcustomtoken';
 const PINECONE_RERANKING = process.env.PINECONE_RERANKING === true || process.env.PINECONE_RERANKING === "true";
 const apiUrl = process.env.API_URL || configGlobal.apiUrl;
 
+let rerankingOff = false;
+if (process.env.RERANKING_OFF && (process.env.RERANKING_OFF === "true" || process.env.RERANKING_OFF === true)) {
+  rerankingOff = true;
+}
+
 
 let MAX_UPLOAD_FILE_SIZE = process.env.MAX_UPLOAD_FILE_SIZE;
 let uploadlimits = undefined;
@@ -1273,6 +1278,10 @@ router.post('/namespace/import/:id', upload.single('uploadFile'), async (req, re
       }
     }
 
+    if (e.situated_context === true && (e.type !== "url" || content.scrape_type === 0)) {
+      content.situated_context = true;
+    }
+
     addingContents.push(content);
   })
 
@@ -1297,8 +1306,6 @@ router.post('/namespace/import/:id', upload.single('uploadFile'), async (req, re
   let engine = ns.engine || default_engine;
   let embedding = aiManager.normalizeEmbedding(ns.embedding);
   let hybrid = ns.hybrid;
-  const situated_context = normalizeSituatedContext();
-
 
   if (process.env.NODE_ENV !== "test") {
     await aiService.deleteNamespace({
@@ -1333,14 +1340,15 @@ router.post('/namespace/import/:id', upload.single('uploadFile'), async (req, re
     return res.status(500).send({ success: false, error: "Unable to get new content" });
   }
 
-  let resources = new_contents.map(({ name, status, __v, createdAt, updatedAt, id_project, ...keepAttrs }) => keepAttrs)
-  resources = resources.map(({ _id, scrape_options, ...rest }) => {
+  let resources = new_contents.map(({ name, status, __v, createdAt, updatedAt, ...keepAttrs }) => keepAttrs)
+  resources = resources.map(({ _id, scrape_options, situated_context, ...rest }) => {
+    const situated_context_obj = aiManager.normalizeSituatedContext(situated_context);
     return {
       id: _id,
       parameters_scrape_type_4: scrape_options,
       embedding: embedding,
       engine: engine,
-      ...(situated_context && { situated_context: situated_context }),
+      ...(situated_context_obj && { situated_context: situated_context_obj }),
       ...rest
     }
   });
@@ -1752,6 +1760,7 @@ router.post('/', async (req, res) => {
 
       const json = {
         id: saved_kb._id,
+        id_project: saved_kb.id_project,
         type: saved_kb.type,
         source: saved_kb.source,
         content: saved_kb.content || "",
@@ -1950,7 +1959,7 @@ router.post('/csv', upload.single('uploadFile'), async (req, res) => {
           situated_context_obj = normalizeSituatedContext(situated_context);
         }
 
-        let resources = result.map(({ name, status, __v, createdAt, updatedAt, id_project, situated_context, ...keepAttrs }) => keepAttrs)
+        let resources = result.map(({ name, status, __v, createdAt, updatedAt, situated_context, ...keepAttrs }) => keepAttrs)
         resources = resources.map(({ _id, ...rest }) => {
           return {
             id: _id,
@@ -2263,6 +2272,7 @@ router.put('/:kb_id', async (req, res) => {
 
   const json = {
     id: updated_content._id,
+    id_project: updated_content.id_project,
     type: updated_content.type,
     source: updated_content.source,
     content: updated_content.content || "",