+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 🚀        IN PRODUCTION                        🚀
 (https://www.npmjs.com/package/@tiledesk/tiledesk-server/v/2.3.77) 
 
+# 2.19.9
+- Fixed an issue with the handling of situated_context during namespace import
+
+# 2.19.8
+- Fix analytics integration
+
 # 2.19.7
 - Added TTL for events collection